+++ test/data.jl
@@ -208,6 +208,14 @@ end
         @test length(cc_ordered) == 2
         @test length(cc_ordered[1]) == 3
         @test length(cc_ordered[2]) == 3
+
+        # arbitrary edge types test
+        data["trans"] = Dict{String,Any}()
+        data["trans"]["1"] = deepcopy(data["branch"]["6"])
+        delete!(data["branch"], "6")
+
+        cc2 = PowerModels.connected_components(data; edges=["branch", "trans"])
+        @test cc2 == cc
     end
 
     @testset "connecected components with propagate topology status" begin